+++ mrblib/ff/logger.rb
@@ -67,10 +67,12 @@ module FF
     #
     # @param [ Object] key Usually the planet id.
     # @param [ String ] msg The none error message.
+    # @param [ Boolean ] compact Join multiple message by whitespace.
+    #                            Defaults to: false
     #
     # @return [ String ]
-    def msg(key, msg = '')
-      errors?(key) ? colorize(key) : msg
+    def msg(key, msg = '', compact = false)
+      errors?(key) ? colorize(key, compact) : msg
     end
 
     private
@@ -78,12 +80,13 @@ module FF
     # Colorized output for the error messages specified by the key.
     #
     # @param [ Object] key Usually the planet id.
+    # @param [ Boolean ] compact Join multiple message by whitespace.
     #
     # @return [ String ]
-    def colorize(key)
+    def colorize(key, compact)
       msgs = errors(key)
       msgs.map! { |e| e.set_color(@color) } if @color != :default
-      msgs.join("\n")
+      msgs.join(compact ? ', ' : "\n")
     end
   end
 end